refactor(tests): Move conftest to libvcs.conf for doctest
